refactor: move comparison functionalities to influencer component
