+++ app/api/backend/influencer/compare/route.ts
@@ -1,10 +1,10 @@
 import { NextRequest, NextResponse } from "next/server";
-import { influencerComparisonController } from "@/interface-adapters/controllers/influencer.comparison.controller";
+import { influencerController } from "@/interface-adapters/controllers/influencer.controller";
 
 export async function GET(request: NextRequest) {
   const searchParams = request.nextUrl.searchParams;
   const usernames = searchParams.get("usernames")?.split(",") || [];
-  const result = await influencerComparisonController.compare({
+  const result = await influencerController.compare({
     searchParams: { usernames },
   });
   return NextResponse.json(result);
