+++ src/interface-adapters/controllers/influencer.controller.ts
@@ -1,7 +1,10 @@
 import { influencerUseCases } from "@/application/use-cases/influencer.use-case";
 import { InfluencerOverviewPresenter } from "@/interface-adapters/presenters/influencer/influencer.overview.presenter";
 import { InfluencerDetailPresenter } from "@/interface-adapters/presenters/influencer/influencer.detail.presenter";
+import { InfluencerComparisonPresenter } from "@/interface-adapters/presenters/influencer/influencer.comparison.presenter";
+
 import { Influencer, Status } from "@/domain/entities/influencer.entity";
+import { getTranslations } from "next-intl/server";
 
 interface IndexProps {
   searchParams: {
@@ -58,6 +61,10 @@ interface LikeProps {
 
 interface UnlikeProps {
   request: { userId: number; influencerId: number };
+}
+
+interface CompareProps {
+  searchParams: { usernames: string[] };
 }
 
 class InfluencerController {
@@ -364,6 +371,59 @@ class InfluencerController {
     const result = await influencerUseCases.unlike(userId, influencerId);
 
     return { pageData: { isSuccess: result.isSuccess } };
+  }
+
+  async compare({ searchParams }: CompareProps): Promise<{
+    pageData: {
+      isSuccess: boolean;
+      influencers:
+        | {
+            username: string;
+            profilePicture: string;
+            averageLikes: string;
+            averageComments: string;
+            averageShares: string;
+            averageSaves: string;
+            averageViews: string;
+            followers: string;
+            status: Status;
+            engagementVisualizationRate: number;
+          }[]
+        | null;
+      error: string | undefined;
+    };
+  }> {
+    const t = await getTranslations("InfluencerComparison");
+
+    const { usernames } = searchParams;
+
+    if (
+      !Array.isArray(usernames) ||
+      usernames.length < 2 ||
+      usernames.length > 5
+    ) {
+      return {
+        pageData: {
+          isSuccess: false,
+          influencers: null,
+          error: t("selectRange"),
+        },
+      };
+    }
+
+    const result = await influencerUseCases.compareStatistics(usernames);
+
+    const influencers = result.influencers.map((influencer) =>
+      InfluencerComparisonPresenter.toHttp(influencer)
+    );
+
+    return {
+      pageData: {
+        isSuccess: result.isSuccess,
+        influencers: influencers,
+        error: result.error,
+      },
+    };
   }
 }
